+++ src/messages/services/banner/index.js
@@ -76,6 +76,7 @@ function mutateMarkup(markup) {
 function fetcher(options) {
     const {
         account,
+        merchantId,
         amount,
         offerType,
         currency,
@@ -97,14 +98,20 @@ function fetcher(options) {
             format: 'HTML',
             presentation_types: 'HTML',
             ch: 'UPSTREAM',
-            call: `__PP.${callbackName}`
+            call: `__PP.${callbackName}`,
+            // Future prep for credit-presentment partner integration, ignored by imadserv
+            merchant_id: merchantId
         };
 
         const queryString = objectEntries(queryParams)
             .filter(([, val]) => val)
             .reduce(
                 (accumulator, [key, val]) => `${accumulator}&${key}=${val}`,
-                stringStartsWith(account, 'client-id') ? `client_id=${account.slice(10)}` : `pub_id=${account}`
+                // TODO: This logic needs to be modified when switching from imadserv
+                // to credit-presentment in order to properly handle partner integrations
+                !merchantId && stringStartsWith(account, 'client-id')
+                    ? `client_id=${account.slice(10)}`
+                    : `pub_id=${merchantId || account}`
             );
         const script = document.createElement('script');
         script.async = true;
@@ -214,7 +221,7 @@ const getContentMinWidth = templateNode => {
     });
 };
 
-const memoFetcher = memoizeOnProps(fetcher, ['account', 'amount', 'offerType', 'countryCode']);
+const memoFetcher = memoizeOnProps(fetcher, ['account', 'merchantId', 'amount', 'offerType', 'countryCode']);
 
 export default function getBannerMarkup({ options, logger }) {
     logger.info(EVENTS.FETCH_START);