+++ tests/unit/spec/src/messages/components/Modal/index.test.js
@@ -1,0 +1,3 @@
+describe('Modal Component', () => {
+    it.todo('Creates the appropriate wrapper');
+});